@@ -1,15 +1,15 @@
 <?php
-/**
- * This file is part of Contao EstateManager.
- *
- * @link      https://www.contao-estatemanager.com/
- * @source    https://github.com/contao-estatemanager/virtual-tour
- * @copyright Copyright (c) 2019  Oveleon GbR (https://www.oveleon.de)
- * @license   https://www.contao-estatemanager.com/lizenzbedingungen.html
- */
 
 declare(strict_types=1);
 
+/*
+ * This file is part of Contao EstateManager.
+ *
+ * @see        https://www.contao-estatemanager.com/
+ * @source     https://github.com/contao-estatemanager/video
+ * @copyright  Copyright (c) 2021 Oveleon GbR (https://www.oveleon.de)
+ * @license    https://www.contao-estatemanager.com/lizenzbedingungen.html
+ */
 
 namespace ContaoEstateManager\Video;
 
